+++ README.md
@@ -1,0 +1,30 @@
+# Going Live
+
+## What is this?
+
+*Going Live* is a super basic containerized starter for easy development of APIs in Go.
+By leveraging the superpowers of the well-know nodemon and the unmatched speed of the Go compiler, the *Going Live* container is able to rebuild the binary and restart the server every time you save some changes or add new files.
+
+## Requirements
+- Docker
+- Docker Compose
+
+## Usage
+```console
+docker-compose up
+```
+
+After changing the Dockerfile
+```console
+docker-compose up --build
+```
+
+Try
+```console
+curl localhost:3001/titles | jq
+curl localhost:3001/titles/1 | jq
+```
+
+## Author
+- [advance467](mailto:advance467@hotmail.com)
+- [alaindet](https://github.com/alaindet)
