+++ cmd/going-live/routes.go
@@ -1,16 +1,13 @@
 package main
 
 import (
-	"github.com/advance467/Going-Live/features/titles"
 	"github.com/go-chi/chi/v5"
+
+	"github.com/advance467/Going-Live/cmd/going-live/titles"
 )
 
 func MainRoutes() *chi.Mux {
-
 	r := chi.NewRouter()
-
 	titles.LoadTitleRoutes(r)
-
 	return r
-
 }
