+++ docker-compose.yaml
@@ -3,8 +3,8 @@ version: '3.8'
 services:
 
   app:
-    build: "./.docker_images/go_live_reloader"
-    image: "go_live_reloader:dev"
+    build: "./.docker_images/going_live_reloader"
+    image: "going_live_reloader:dev"
     user: "node"
     working_dir: /home/node/app
     volumes:
@@ -17,20 +17,26 @@ services:
       - database
 
   database:
-    image: 'postgres:15.1'
+    image: 'postgres:15.1-alpine'
     ports:
-        - '5433:5432'
+      - '5433:5432'
     environment:
-        POSTGRES_DB: 'goinglive'
-        POSTGRES_USER: 'goinglive'
-        POSTGRES_PASSWORD: 'goinglive'
+      POSTGRES_DB: 'goinglive'
+      POSTGRES_USER: 'goinglive'
+      POSTGRES_PASSWORD: 'goinglive'
     volumes:
-        - 'going-live-db1:/var/lib/postgresql/data'
+      - 'goinglivevol:/var/lib/postgresql/data'
+      - ./database/pg-init.sql:/docker-entrypoint-initdb.d/pg-init.sql
     healthcheck:
-        test: ["CMD", "pg_isready", "-q", "-d", "goinglive", "-U", "goinglive"]
-        retries: 3
-        timeout: 5s
+      test: ["CMD", "pg_isready", "-q", "-d", "goinglive", "-U", "goinglive"]
+      retries: 3
+      timeout: 5s
+
+  adminer:
+    image: adminer:4.8.1
+    ports:
+      - '8081:8080'
 
 volumes:
-    going-live-db1:
-      driver: local
+  goinglivevol:
+    driver: local
