--- main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"net/http"
-)
-
-func main() {
-
-	http.ListenAndServe(":3000", MainRoutes())
-
-}
